+++ openjaus/trunk/OpenJAUSv3.3/libjaus/src/message/inform/manipulator/reportJointPositionsMessage.c
@@ -160,7 +160,7 @@ static int dataToString(ReportJointPositionsMessage message, char **buf)
     jausDoubleToString(message->jointPosition[joint], (*buf)+strlen(*buf));
   }
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
 }
 
 // Returns number of bytes put into the buffer
@@ -560,6 +560,6 @@ static int headerToString(ReportJointPositionsMessage message, char **buf)
   strcat((*buf), "\nSequence Number: ");
   jausUnsignedShortToString(message->sequenceNumber, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
   
 }
