+++ openjaus/trunk/OpenJAUSv3.3/libjaus/src/message/inform/worldModel/reportVksDataTransferTerminationMessage.c
@@ -489,6 +489,6 @@ static int headerToString(ReportVksDataTransferTerminationMessage message, char 
   strcat((*buf), "\nSequence Number: ");
   jausUnsignedShortToString(message->sequenceNumber, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
   
 }
