+++ openjaus/trunk/OpenJAUSv3.3/libjaus/src/message/query/manipulator/queryJointVelocitiesMessage.c
@@ -493,6 +493,6 @@ static int headerToString(QueryJointVelocitiesMessage message, char **buf)
   strcat((*buf), "\nSequence Number: ");
   jausUnsignedShortToString(message->sequenceNumber, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
   
 }
